Adding in initial code
The code here is based on an earlier version of the same tool,
and we have not yet done the necessary work to clean it up and
make it correct for our purposes. Instead, it is more important
that we have *something* we can invoke for the tea project, and
then iterate.